Pylinted verdict.py and moved classes between it and basics.py more logically
