+++ tcsfw/android_manifest_scan.py
@@ -10,7 +10,7 @@ from tcsfw.model import IoTSystem, NodeComponent
 from tcsfw.property import Properties, PropertyKey
 from tcsfw.tools import ComponentCheckTool
 from tcsfw.traffic import EvidenceSource, Evidence
-from tcsfw.basics import Verdict
+from tcsfw.verdict import Verdict
 
 
 class AndroidManifestScan(ComponentCheckTool):
