+++ tcsfw/zed_reader.py
@@ -11,7 +11,7 @@ from tcsfw.model import IoTSystem
 from tcsfw.property import Properties, PropertyKey
 from tcsfw.tools import BaseFileCheckTool
 from tcsfw.traffic import EvidenceSource, Evidence
-from tcsfw.basics import Verdict
+from tcsfw.verdict import Verdict
 
 
 class ZEDReader(BaseFileCheckTool):
