+++ tests/test_dhcp.py
@@ -1,6 +1,6 @@
 import pathlib
 from tcsfw.address import HWAddress, IPAddress
-from tcsfw.basics import Verdict
+from tcsfw.verdict import Verdict
 from tcsfw.builder_backend import SystemBackend
 from tcsfw.event_logger import EventLogger
 from tcsfw.inspector import Inspector
@@ -8,7 +8,7 @@ from tcsfw.main import DHCP, UDP
 from tcsfw.matcher import SystemMatcher
 from tcsfw.pcap_reader import PCAPReader
 from tcsfw.traffic import IPFlow
-from tcsfw.verdict import Status
+from tcsfw.basics import Status
 
 
 def test_dhcp():
