+++ tests/test_event.py
@@ -1,5 +1,5 @@
 from tcsfw.address import Addresses, EndpointAddress, Protocol
-from tcsfw.basics import Verdict
+from tcsfw.verdict import Verdict
 from tcsfw.builder_backend import SystemBackend
 from tcsfw.event_interface import PropertyAddressEvent, PropertyEvent
 from tcsfw.main import DNS
