+++ tests/test_model.py
@@ -1,11 +1,11 @@
 from tcsfw.address import EndpointAddress, Protocol, DNSName, IPAddress, HWAddress
-from tcsfw.basics import Verdict
+from tcsfw.verdict import Verdict
 from tcsfw.builder_backend import SystemBackend
 from tcsfw.main import UDP, SSH
 from tcsfw.matcher import SystemMatcher
 from tcsfw.basics import ExternalActivity
 from tcsfw.traffic import IPFlow
-from tcsfw.verdict import Status
+from tcsfw.basics import Status
 
 
 def simple_setup_1(external=False) -> SystemBackend:
